+++ NEWS.md
@@ -1,6 +1,7 @@
 # Unreleased version
 
 - Snippets: when inserting code snippets, `library()` calls should not be removed (in other LCA wizards, duplicate `library()` calls are removed)  
+- Transformations: When input data has over 1000 rows, show a warning and only use first 10k rows for performance reasons
 - Transformations: Previously, the transformations code was getting executed in the global environment after each transformation was made, which meant that new variables were getting created in the global env even without clicking the "Insert Code" button. Now the code only runs after choosing to "Insert Code".
 - Transformations: fix bug where filtering a column that's an ordered factor (for example, "cut" or "color" in the `ggplot2::diamonds` dataset) was resulting in an error
 - Transformations: fix bug where missing values in numeric columns were not treated as missing values